Rename package to match lambda and directory
